# Speed up coh_tmm in tmm_core_vec; introduced versions with cpu and gpu parallelization