… solcore/absorption_calculator/transfer_matrix.py
… solcore/material_system/material_system.py
… solcore/structure.py