s and p polarization, 10000 wavelength x angles, 6 layers, calculate coh_tmm:
before speed increase: 4.178s
after speed increase: 2.399s
after speed increase, non-detailed mode: 1.948s
CPU parallelization with 24 cores
CPU parallelization: 0.844s
CPU parallelization, non-detailed mode: 0.719s
GPU parallelization with NVIDIA GeForce RTX 4060
GPU parallelization: 0.296s
GPU parallelization, non-detailed mode: 0.118s