@phoebe-p 2024-05-07
sorry for my lack of input on this. I was wondering, for the parallelisation, what do you think the best way to incorporate this would be? I guess there are now three options: no parallelisation (i.e. just the old implementation), GPU parallelisation, or CPU parallelisation. Then there's the detailed vs. non-detailed mode. The user should be able to choose which one they want to use, but it would be better not to have three different files tmm_core_vec files, since most of the content is the same anyway.